# fix: improves checkOwnerRef func perf